requirements.txt
Click==7.0
Flask==1.1.1
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.17.0
opencv-python==4.1.0.25
Pillow==6.1.0
pytesseract==0.3.0
tesseract==0.1.3
Werkzeug==0.15.5
